Update README and documentation landing page
Reduce the content of the README so that it does not easily get outdated. Rewrite the landing page, taking Spark as an example, to be more direct and concise. Remove the introduction for the developer documentation, as it is verbose and confusing, clarify also the developer documentation landing page.
Move the Kubernetes backend page to the top level as it is about setup, not development.