Implement image-aware scheduling and pre-loading
This feature is implemented only on the DockerEnging back-end.
When enabled via the new configuration option --backend-image-management, Zoe will (1) when an execution is submitted, start threads to pull the image on the nodes and (2) schedule services on the hosts that have the image already loaded.
This has two big side-effects:
1. Zoe will not let Docker automatically pull the image at run time and will refuse to start a ZApp with unavailable images
2. ZApps will always run with the latest images: a pull is always done before a run